# wwwdir/venom/shared/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Api
  description: VENOM CODES API specification — WHMCS-backed client portal
  version: 0.1.0
servers:
  - url: /
paths:
...
  /api/products/{productId}:
    get:
      tags: [Catalog]
      summary: Single product
      operationId: getProduct
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyJson"
...
  /api/paymethods/capture:
    post:
      tags: [Billing]
      summary: WHMCS CapturePayment (invoiceid)
      operationId: capturePayment
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [invoiceId]
              properties:
                invoiceId:
                  type: string
                cvv:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyJson"
...
  /api/oauth/token:
    post:
      tags: [Auth]
      summary: WHMCS CreateSsoToken (SSO), not OAuth2 token URL
      operationId: createSsoToken
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                destination:
                  type: string
                serviceId:
                  type: string
                domainId:
                  type: string
                userId:
                  type: string
                ssoRedirectPath:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyJson"
  /api/invoices/{invoiceId}/pay:
    post:
      tags: [Billing]
      operationId: payInvoice
      security: [{ bearerAuth: [] }]
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyJson"
...
  /api/paymethods:
    get:
      tags: [Billing]
      operationId: getPayMethods
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PayMethodCard"
    post:
      tags: [Billing]
      operationId: addPayMethod
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddPayMethodBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /api/paymethods/{payMethodId}:
    delete:
      tags: [Billing]
      operationId: deletePayMethod
      security: [{ bearerAuth: [] }]
      parameters:
        - name: payMethodId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyJson"
...
  /api/products/{productId}/addons:
    get:
      tags: [Catalog]
      operationId: getProductAddons
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductAddonsResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
    AnyJson:
      type: object
      additionalProperties: true
...
    PayMethodCard:
      type: object
      required: [id]
      properties:
        id:
          type: string
        type:
          type: string
        gateway:
          type: string
        description:
          type: string
        lastFour:
          type: string
        expiry:
          type: string
    AddPayMethodBody:
      type: object
      required: [gateway, description]
      properties:
        gateway:
          type: string
        description:
          type: string
        token:
          type: string
...
    ProductAddon:
      type: object
      required: [id, name, pricing]
      properties:
        id:
          type: string
        name:
          type: string
        pricing:
          type: object
          additionalProperties: true
    ProductAddonsResponse:
      type: object
      required: [addons]
      properties:
        addons:
          type: array
          items:
            $ref: "#/components/schemas/ProductAddon"
...